Se cambiaron las id de Integer a Long

+++ src/main/java/com/quantumzone/QZ_Workhub/persistencia/entidad/Notificacion.java
@@ -15,7 +15,7 @@ public class Notificacion {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_notificacion", nullable = false)
-    private Integer idNotificacion;
+    private Long idNotificacion;
 
     @Column(name = "motivo", length = 45, nullable = false)
     private String motivo;
