+++ src/main/java/com/quantumzone/QZ_Workhub/persistencia/entidad/Sala.java
@@ -15,7 +15,7 @@ public class Sala {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_sala", nullable = false)
-    private Integer idSala;
+    private Long idSala;
 
     @Column(name = "nombre", length = 45, nullable = false)
     private String nombre;
